Add tests for URLContentExtractor
I added tests for read and sanitize methods

Added a bunch of tests for the sanitize method for many different
inputs that may come up. Also added an if statement for contents ==
null in URLContentExtractor.sanitize so that it wouldn’t throw an
exception or break.

I added two different tests for read, one that pulls from the internet
(i used example.com which doesn’t normally change) which could break if
the site changes and one with mockitto which is a stupid test because
we are just asserting that a constant is equal to itself. I don’t
really like either test. Not sure if there is a better way to test this
in isolation.

I also added a readAndSanitize method so that we could write a test
that actually shows what this part of the project is doing, and mocking
out the read request there makes a lot more sense.