Tidying up the main
Moved the input and output to their own class files. Added a method for
sorting Map values. Changed the custom word list to take a file path
anywhere on the system instead of just the name of a file in the
resources folder. Changed a MarkUpTest test to correctly test for String
equivalency.